Added s_encode and s_decode, improved evaluate, reduce, and simplify in Kolmo.SK
Added a boolean argument to avoid sorting a tree during the printing using
AsciiTree.draw_parenthesis and a conditional to avoid printing empty [].

Added s_encode, s_decode, added some builtin functions to the SK engine and the
sk_define to define macros for quick definition of "functions" in SK.

The default  behavior of simplifying every expression can be turned off using an
optional argument (tail_too). This is useful when applying the Fixedpoint macro
for recursive calls that require lazy evaluation.