Removed string representations and added sequence functions.
Modified the interface of most functions to avoid the use of strings whenever
that's possible, to reduce computation time of complex expressions. This caused
changes in the representation of the internal trees (Not Parser.Tree.t anymore),
and created the type sk.

Added Substitution, Deletion, and Insertions as function examples.

Fixed a bug in the Decode function that rendered it useless.