Added support for the SK DSL in the Kolmo.SK.Parser module.
Improvements:

The Kolmo.SK.Parser module contains a Camlp4 specification to parse a simple SK
definition language. It supports in its current form only

def: to define expressions
eval: to evaluate an expression and print it's result

Examples of this language are:

def A B C D = (B (C D));;
eval (A X Y Z);;

To activate it, it is necessary to run Kolmo.SK.Parser.activate () from
./mpoy_console kolmo.cmo (or your own program during initialization).

Interface Changes:

Delayed the expansion of macros to the evaluation step in an explicit
call. This improves readability in most cases.

Added Kolmo.SK.sk_define_interpreted to define functions from a preparsed tree.