Improved the behavior of the GenBank file format parser.
The GenBank format parser used to take the GenBank file and reconstruct all the
information contained in the Fasta format (as produced by genbank). This had the
tremendous disadvantage that the terminal names where useless in a tree. To fix
this, we now extract the organism name, and replace the spaces with underscores.

Much better.