Improved detection of Hennig/NONA/TNT files.
The Hennig/NONA/TNT files where only detected if the first command is xread or
nstates. Moreover, the command should have occurred in the first line of the
file. Mesquite produces files that start with the quote command, and therefore,
POY failed to detect the format. I have modified the check to evaluate all the
lines of the file looking for the xread and the tread commands. This yields a
more solid file type detection for Hennig/NONA/TNT.