Improved error messages for Mesquite's TITLE command in NEXUS files.
Mesquite have added a new command to the TAXA block: the TITLE command. This
command is NOT a valid NEXUS command. One wonders why the authors of the file
format would add yet another non-valid modification. Anyway, POY does not
recognize this command. instead of giving a meaningless error to the end
users, this change generates an error message explaining the user the cause of
the error and where to look for information to correct it (the mesquite
website).