Modified the default algorithm of build (_mst).
The default algorithm of build (_mst) used a depth first search starting with
the worst neighbor. I have modified it to use a dfs using the closest neighbor.
To implement it I had to modify one constructor from Closest to Closest2 to
eliminate some name clashes with another Mst type.