Improvement: Removed duplicated tree evaluations during TBR.
During TBR, the new swapper was calling a full SPR round after each
reroot.  However, this is unnecessary, as trees are evaluated more
than once in this scheme. It is only necessary to apply one side
of SPR, by connecting as child of the rerooted subtree to the
original parent subtree.