Simplified the automated search function and improved performance.
The automated search function now chooses better the algorithm depending on
whether or not there is dynamic homology data. Additionally, ratchet is now
applied by selecting a tree from the pool of best trees, not only on the current
tree just built. This yields more potential for finding better trees (and it
does of course).