Removed groff from the configure script and improved man page.
The man page used to be formatter by the make rule using groff. This added an
unnecessary requirement to the compilation process, and made the man pages
clumsier. Now we store directly the man page with the troff commands, which
makes it look better in all systems.

Also added the authors to the man pages.