Added nexus POY block for branch lengths of multiple character sets, + 2 issues
Two features of the nexus specification were added. Translations, in the TREE
block can also be an ordered list of taxa which is associated with the numbers
1 .. n. Also in this block, the UTREE keyword can read preceeding a tree, as in
the example below.

With likelihood and multiple character sets we needed a way to have branch
lengths associated with each character,tree, and node. This is done using the
comments (which should be ignored by other applications) and the addition of a
POY block.

Begin POY;
  CharacterBranch
      [Names = <character-set-name>[,<character-set-name>...];]
      [Trees = <tree-name>[,<tree-name>...];]
      Map <node-name length>[,<node-name length>...];
  ;
END;

character-set-name are defined in the BEGIN SETS; ... END; block, and trees are
marked with the trees. node-names are defined in comments, preceeded by &P after
a node, or it's branch length. If branch lengths are also provided they are
overwritten by the poy blocks interpretation.

Example (matrix and header excluded):

BEGIN SETS;
    CharSet ch1 = 1-200
    CharSet ch2 = 201-500
END;

BEGIN TREES;
    translate alpha,beta,gamma;
    utree one = ((1:0.02[&Pnode1],2[&Pnode2])[length unknown],3[&Pnode3]);
END;

BEGIN POY;
    CharacterBranch
        Trees = one;
        Names = ch1;
        Map node1 0.123, node2 0.0987, node3 0.3419;
    ;
    CharacterBranch
        Names = ch2;
        Map node1 0.161, node2 0.1987, node3 0.0419;
    ;
END;