Added the sankoffMatrix.ml script to create sankoff characters.
To use:

cd src
make mpoy_console
./mpoy_console ../utils/sankoffMatrix.ml input_script.poy

The input_script.poy contains a poy script that only reads the
characters of interest and sets their parametres. The output is a
dpread file suitable for poy use, named input_script.poy.ss (the name
matches that of the input selected by the user).