If a compressed file is going to be read, then close the outchannel and flush.
A potential source of problems could be users doing:

swap (visited:"bleh")
report (supports:bremer:"bleh")

As the file bleh is not closed, then it does not have the checksums expected at
the end of the file. To remove this problem, when a compressed file is read, we
first check if the channel is opened, flush and close it if it is.