Added decoder trees for shorter function definition in kolmoExtension.
In the previous versions of kolmoExtension, one could only define functions in
the form:

let a b c = pair b c

which would inline the (previous) definition of pair in the function a. I have
now added the function constructor:

val pair a b c = c b a

val apair b c = pair b c

which effectively assigns a code to pair, and such code is used in the call of
pair inside a, while assigning apair a code on its own. The compiler takes care
of automatically adding a decoder to the arguments for callbacks within each
function (so that the code of pair can be decoded and the correct function
definition used).