Made the old poyParser grammar for Kolmogorov characters a function.
The old grammar for Kolmo characters was compiled and added to Camlp4 at
initialization time. Because of the way it was written, this made it difficult
to use the same keywords in other grammars, in particular, the new Kolmogorov
character grammar which uses fundamentally the same keywords.

To eliminate this problem, the grammar will only be generated if used. This is a
temporary change as the old grammar will eventually be eliminated.