Connected the new module Kolmo to the Data module and removed PoyParser.
The functionality of PoyParser was deprecated. The only non-Kolmogorov
functions that it handled where explode_file and guessing the file type and
immediately adding its contents to Data.d. These two functions where moved to
more natural places: Parser and Data.

All the Kolmogorov functionality will be replaced with the SK machine
implemented in Kolmo.