Second functional version of Kolmogorov characters. New alignment function.
I have modified the structure of the Kolmogorov characters. Due to the
complexity of the operations involved, and the level of specialization, as well
as the amount of experimentation that will be required, I have packed them
inside the kolmo directory and use the ocaml -pack functions to produce the
corresponding module.

This module now packs an alignmet function that provides a specialized version
intended only for Kolmogorov characters, where the distance from previous
evolutionary events is included in the overall alignment cost.