Added support to create trees with only one or no leaves.
POY used to fail if the input data had only one leaf. I need this
feature for Kolmogorov dataset tests. To support it, I created a
simple wagner tree manager that only took the initial tree and
returned it. The absence of this manager was the only reason for
not supporting this kind of trees.