Various corrections in the total cost of a node. Needed for Kolmogorov.
The total cost of the leaves in Kolmogorov characters is not 0. For this reason,
a number of functions needed to initialize the leaves by adding the cost of the
Kolmogorov characters.

Moreover, a bug in the root cost assigned by to_single (that Nick corrected but
has not pushed), has delayed my work (he can't really push that change without
breaking everything!). So I have fixed the bug and he will have to deal with the
merge of those lines.