Bug fixes and rewrite of uppass heuristic
The major improvement here are the additions of assertions, the rewrite of the
uppass heuristic function and it's callee (refresh_all_edges), and when and how
it gets called.

The uppass heurisitic transforms adjusted and unadjusted with branch lengths in
alldirections. An assertion to verify branch lengths in all directions is
called. Assign single places a root at the location that would denote how the
tree should be traversed to get the assigned single nodes.

Along with this the break and join functions use refresh all edges where needed,
such as applying the three directions after a single direction is applied to
it's nodes from inserting/removing internal nodes.

Other Changes -

Bug Fix in to_single in node.ml
   - cost was being calculated from median of two nodes, not a choice of one of
     the nodes. A median is only required for likelihood characters.

Change of Signature for uppass_heuristic -
   - default branch lengths are not required here, since they are applied to the
     tree in the downpass. This simplifies the usage of the uppass heurisitc so
     the times are already in the tree and just need to be transfered to a nodes
     neighbors.

Change of Signature for get_times_between -
   - the second argument is a node option (first being a node), this is for
     roots when we want to add the two branch lengths together to apply that
     total time to each of the children (we don't care about the children, so
     the second argument, 'child', is unnecessary).