check_cost now follow a postorder traversal for the edge distance summation.
The check_cost function needs to add the branch lengths to compute certain tree
costs. However, the order is of relevance for certain computations, in
particular Kolmogorov characters, and maintainin the invariants for the
three-direction characters (changes that Nick will merge).

To correct this, the addition of the edge length is now done using the
Tree.post_order_node_with_edge_visit_simple function.  Note that this function
adds the root cost twice, so we remove the direction that we are not interested
on in its the accumulator.