iteration during a build was not iterating completely
Cause - The set of nodes to iterate after a join (based on the deltas) were
not completing all the nodes that should be updated, since, now, the nodes that
get iterated in the tree are done deterministically (post order).

Solution - redo all the nodes for now, but this can be limited to fewer later.