Better likelihood iteration and initial estimation
Using Brent's method of Iteration, a combination of golden section search and
parabolic interpolation. Currently in pure OCAML only.

Added estimation function for JC69 which is,
       d = - 3/4 * log ( 1 - 4/3 * p )
where p is the proportion of site differences. To add better estimation for
other models, although, how to resolve this between two internal nodes is an
issue.