Removed references in Kolmo.Compiler.
The Kolmo.Compiler used internal references to keep the state of the compiler.
That did not allow a nice integration with POY for Kolmo-Complexity analysis. I
removed now the references to make is easier to plug the new compiler in the
rest of the application. This is necessary to be able to support the on-the-fly
transforms between different character types.