Added the Data.d structure to the contents of Ptree.p_tree.
The adjustement of parameters in Likelihood and Kolmogorov characters requires a
per-tree set of model parameters. A long needed change then was to include the
data associated with a tree in the tree itself. This also simplifies a number of
interfaces (that have not been updated yet).

Nevertheless, this change is the first step in the process of adding iterative
parameter adjustement.