Fixed failure in AllDirNode.AllDirF.is_collapsable.
The is_collapsable function assumed that the adjusted assignment to the vertex
was a unique element. This need not to be the case after the Likelihood related
updates.

Therefore, we take the first element as the adjusted item, as the implementation
guarantees that this is the adjusted element when more than one element in the
list exits (Nick pers. comm.).