Removed the character scripting modules.
The character scripting modules (CharacterScripting and
AllDirChar.CharScripting) where really dead code modules. Their initial goal was
to provide the users with functionality to perform operations in the characters
themselves directly. This functionality was never really added, and the modules
only bloated the code.

Now that I'm updating the Scripting.run record, it is time to do some cleanup too.