Iterate one of the two branches during the readjust function in likelihood
This is part of a larger set of updates that will pull together the iteration of
the tree (described below). This update only iterates the minimum branch of the
subtree -- to get deterministic results. It will be changed to a random choice
once testing is complete. Following is the description of the full method:

requirements of method:
   1 - able to iterate branches in different directions
   2 - only iterate one of the branches of an internal node of the rooted tree.
       This is due to the pully principle.

as a result, we will doing the following:
   folding over edges connecting to a leaf node do
        reroot at this edge
        iterate random side of each node in the tree delta from reroot
        if delta < E then break, else continue
   done