Removed function that mutates Data.d in ptree to functional style.
The branch length table is only used to preserve the rediagnosis function --so
it will return a tree with the same branch lengths and likelihood. So, when any
other operation is performed (iterative, join, break, et cetera), the branch
lengths in the tree should be cleared. The clear_internals function was used
since the name fits and the function isn't used, anywhere.