Added commands to estimate likelihood model based on current tree
--
The model is estimated based on the transitions of nonaddative characters, the
priors on the leaves, and the number of invariant columns. Further
building/reading of trees is started with baseline models of the parameters for
example, GTR has all ones, and K80 with 4.0. The new command is exactly the same
as the likelihood command, but it will estimate the transitions.
    elikelihood(model,variation,priors,gaps)

--
The specification for a model was moved into the model itself. This made it
easier to update the model and specification at the same time for consistency.

--
Functions for counting the transitions and converting between alphabets was made
in the nonadditive character modules.