Added search functions for model, and an initial estimation.
The estimation uses nonaddative characters and a tree to determine the number of
transitions between states in the tree. This routine is seperated from other
transformations since it requires and is dependent on a tree. No tree results in
initial values being estimated -- see MlModel.ml for initial values.

The model search function uses the BFGS method, and a line search routine, to
minimize the likelihood of the _entire_ tree. Searching with a subset of
branches and characters will eventually be possible with added search strategies
for the iterating the model during a tree search is performed.

Bugfix:
    Symptom - using_likelihood would always return false.
    Cause   - ifdef was, USING_LIKELIHOOD not USE_LIKELIHOOD.
    Solution- Changed to correct spelling.

Bugfix:
    Symptom - None; ran into situation when adding estimation procedure.
    Cause   - Data.d, being individual to trees now, was being updated globally
              and not specific to the tree.
    Solution- This started moving the estimation procedure to a tree
              transformation, as mentioned above, and ensuring the replace
              function was only being called on copied Data hashtables.