bug fix for "repeated int code in median sequence" for siepel median solver
Symptom: the result of siepel median solver is an array of int, it should be
standardized, like "1,2,-3,4,6,-7..". But we got repeated items in the array
sometimes.

Cause: function "copy_with_reversal" in med_utl.c is generating repeated items,
because the input pointer "dest" and "src" are the same.
The cause of this is the memory block "mm->vf" we are using -- a vertex factory,
provide empty vertex, which will be filled with possible permutation of
a candidate sequence. If the result of filling is good, we push the vertex
to candidate pool "mm->ps".

Solution: mark the vertex as "inuse" when we push it to "mm->ps". Unmark it
after we kick it out of the candidate pool. In function "get_vertex", when
current head vertex is "inuse", try the next one, till we reach an "unused" one.