Added an iteration manger to build and swap routines; and user-level commands.
The iteration manager defines what edges to adjust and when to optimize the
model during building and swapping. This is done though objects in the
nodes_manager. As such, set(iterative..) does not need to be set for these
commands to work properly.

The current iteration patterns are as follows:
    for the model:
        always      -- iterate on every join
        never       -- no iteration is done
        threshold:X -- only iterate the model if the score is within a certain
                       amount (X) of the starting score.
        max_count:X -- iterate every X joins.
    for the branches:
        all         -- iterate all the branches at every join.
        join_region -- the region surrounding a join (maximum 5 branches).
        join_delta  -- the path from the break to the join point.
        never       -- do not iterate branches of the tree.

These options can be specified as arguments within the build and swap commands
using the iterate argument. For example,

    build( iterate:(branch:all, model:(max_count:5)) )