modified the point_normal function to reduce complexity
The three equations within the body of the function only differ by the
coefficients used in the equation. So, the equation was removed from the
function body and an external function takes care of it; being passed arrays of
the coefficients.

This should increase comprehension of code as it will make it obvious that the
equations only differ by their coefficients.