Removed data requirement in reporting diagnosis
An alphabet was added to the model, and that should be used to report the
likelihood data, and doesn't need to be reconfirmed with the Data.d table.

This helps when reporting the formatter for dynamic likelihood characters when
the data doesn't match with the static characters --it would when utilizing the
dynamic_static_codes and static_dynamic_codes maps they would match.