added simple utility to compare trees using partitions; and branch lengths
usage ./tree_compare.### treefiles

accepts an unlimited number of tree files, each can contain any number of trees.
Each tree will be compared to THE FIRST TREE, and print out the partitions that
are different from this tree. If the tree is the same, then a comparison of the
branch lengths will be done.