bugfix; categorizing after syncing data
We need to categorize the new data into the proper lists, especially if we are
transforming to different character sets. These changes reflect the differences
in the dynamic likelihood section, moving between static and dynamic characters
during the uppass and swaps.

also, small cosmetic issues in node.ml, and mlModel.ml