Stabilizing Priors during dynamic likelihood
A consistency issue came about when we are in dynamic likelihood. We start off
with unknown priors, and do a _very_ convervative estimate (the sum of the longest
sequences minus the length each sequence, the minimum number of additional gaps
required to bring all the sequences to the same length). This proved fine, but
once an implied alignment is known, then new priors based on that alignment are
found. Now we have an inconsistency. The priors used in the downpass of the
static characters is different then the priors of the alignment. Thus, this
stabilization addition continually aligns and estimates priors until something
converges. Of course, we have noticed osscilations in the convergence, thus we
max out at 10 passes before we leave the algorithm.