Added MPL cost function for median of three elements; and adding 1
Added median1 and median3. Median1 essentially adds data to a median2, while median3
would calculate the information de neuvo. This is done because the MPL
characters do not behave well with a branch length of 0; cost is defined on the
internal nodes of the tree. Median1 is used on roots, and should not be called
under MAL cost mode. This functionality can be added later using median2 with
0.0 branch lengths, just to fill int the gaps.