Clarified some details regarding Coupled/Independent in likelihood
Coupled and Independent made sense under GTR, but had confusing interpretations
under simpler models. Under GTR the rate parameters for gap are coupled, it
means they are all the same, and independent, each is optimized independently of
each other. For simpler models, Independent had meant that the gap rate was
the same as the base rate, now, that is coupled, and independent adds a rate
parameter. In general Independent adds more rate parameters to the model then
Coupled.