Added MPL Alignment; factored out commonality in MlDynamicCS
The dynamic mpl alignment code assigns a state to the median that represents the
best cost of transforming the left branch of that length, and independently, the
right branch from another length.

To do this, with a unified interface for FP and MPL alignment, the two branch
lengths are carried over until the last minute (when creating the cost function)
for the FP alignment code.

Caveat; the median for the MPL alignment needs to be completed by using the
median states assigned to the matrix, and not one/both of the ancestor
sequences; that, and further testing will be part of another changeset.