Added ability to estimate priors from nexus files;
command in nexus:   priors = estimate

Will estimate priors based on the character blocks defined by the likelihood
character set. There was some workings that needed to be gotten around, noteably
calculating the frequencies of characters then passing the results to a general
function in MlModel. This should help unify the process of calculating the
priors, even though the method itself isn't derived from that module.