Added Robinson Foulds distance to tree module
Formally a part of tree_compare, in util, this distance function measures the
number of operations to a median tree by compressing and expanding edges. This
essentially amounts to finding the number of partitions that are the same from
one tree in another, and vice versa, and summing the results.