Script to aid in bisecting the repo;
This script will aid the user in running bisect with POY and a script. The
script is interactive (you need to tell the script on each iteration if the
bisection passed or failed or it should be skipped). Although not 100%
automated, it does take away the tediousness of re-compiling and errors in
missingsteps in the compilation process.

The script is run as,

./bisect.sh <POYSCRIPT> <LOWERBOUND> <UPPERBOUND>

POYSCRIPT - the poy script to run; ensure that it has an exit () command to move
            things more smoothly, although it is not necessary.
LOWERBOUND- (optional) the lowerbound is going to be the last known GOOD
            changeset. the default is 773. required argument if the UPPERBOUND
            needs to be specified.
UPPERBOUND- (optional) the upperbound is the last bad revision. This argument
            defaults to the current revision.