bugfix; in estimating priors. improvement; better specification of rates
symptom - index out of bounds error
cause   - did not subtract one from array length
solution- subtract one from array length

improvement; the command to specify rates in the likelihood command becomes,

 rates:(gamma/theta/none/constant)

This is to allow susinct specification of the model, and remove ambiguity on
what gamma/constant/ or some other option is doing. We know that they are
related to the rate varation/classes.