Warning Message in Nexus; remove extra output; better error message
The warning message in nexus files arises when static character data of
different lengths is read into the data matrix. We detect no space characters
preceeding the name of a taxon. This should indicate that the last call to
consume data ended in the middle of a word and there could potentially be
something wrong. We issue a warning and continue. Later checks for trees or data
can help confirm such errors and error out as per usual.