Modified the diagnosis output for dynamic likelihood
There were some problems with how the output was being displayed for dynamic
likelihood characters, and the following have been resolved.
    1) under MPL/FLK we can resolve the sequence data under the States column
    2) no branch lengths were being displayed to terminal
    3) naming conventions for the XML were inconsistent
    4) navigating XML was difficult --find_args/find_tag have been added

One thing that does not get printed during the output is the model information.
This is difficult to perform, since each class/character can have a different
model. Should an additional column be created to resolve this? Or should this be
left for more verbose output? My feelings, since there is a report(lkmodel)
command, is to leave it for more verbose output, or make a speciallized column
with a short name of the model.