bugfix for uppass under dynamic likelihood re-estimating branches
Symptom - noticed calls during debugging for something else that edges were
          being estimated again during uppass.
Cause   - this information was not being passed to the distance_of_type
          functions.
Solution- Add this information to the function.
Caveat  - Do we need this in the other distance function?

Added the cost calculation for single cost in dynamic likelihood.