Bugfix; parser including '=' in IDENT. Improved some code/changed type.
symptom - reading a tree like, "TREE name= ((1,2),3);" failed
cause   - = is part of IDENT and being consumed by the previous token. A equal
          sign is required, a name is optional, so this can cause lots of
          problems.
solution- exclude '=' from the lexer of IDENT in the tree parser.
Caveat  - should Ident be the same in the tree and nexus body parser?


Also, cleared up the type for charset by changing the strings to int. They are
used as ints, matched as ints, they should be used as such instead of converted
to and from strings.