Initial check-in for custom models in likelihood
Custom models are ones were we link sites in the matrix together, so we can
emulate all other models. This is done by reading in a matrix of characters that
associate the transformations,

ex,
        - a a a a
        a - b b b
        a b - a a
        a b a - b
        a b a b -

The matrix is checked for being symmetric, and diagonal characters are ignored
and normalized based on the other states --as usual. The matrix is multiplied
through by the priors (if one didn't want this priors should be set to equal).
Optimization can be done on, and they are initially set to 1.0 (jc69), as GTR
and other models are.