Consolidated start-up to ensure arguments are processed first
Issue - output.xml was being created no matter the command line arguments

Solution - All output commands (being part of starting up the application), are
moved into poy.ml; this allows other modules to be loaded (as a dependency of
poy.ml), before the arguments are processed and certain actions taken place.

This is something that needs to be looked at further for other command line
arguments -- for example, we could initialize a matrix randomly, but init will
not be initialized yet since poy.ml hasn't. This is something that cannot be
monitored by the compiler, and must be a conscience effort by the programmer.