Fix for gcc-4.5+ and the link order imposed by ocamlbuild
the link order places external libraries before the libraries that ocamlbuild
previously built --even though some of those ocamlbuild built libraries need
these external ones. Link order is stricter in 4.5+, and these command line
arguments over-ride the newer versions. This could be depreciated in the future,
and is absolutely a workaround for ocamlbuild.