Added General Additive characters for ranges > 0 -- 255. Removed dead code.
Identified that AddCS only needed a few external functions to replace the
previous vectorized code that allowed only 256 states. A new module AddGen was
created to replace the external functions (which were moved to AddVec). This
functorization is not complete, and the AddVec module is included in the module
to allow the functorization to be transparent for the time being.

The choice, between General or Vectorized is automatric and data could even be
demarcated into two sets to allow further speed increases. These changes will
proceede in future releases.