Added Numerical.simplex and Numerical.subplex Optimization routines
These routines are proven to work better at noisy function optimizations then
other derivative routines --why take a non-sensical derivative? The functions
need some work to abstract all the details of any optimization routine and to
abstract the processes into a unified set of strategies that can be chosen by
the user based 1) performance 2) data 3) accuracy of the algorithm results.

Also included is the first iteration of a numerical testing application that
runs four metbods (brent, bfgs, simplex, subplex) and determines the mean and
standard deviation. The application comes with a number of functions oft used in
testing numerical optimization routines. Adding further in the future, as well
as a meta-poy to diagnose trees will come soon.