Bug Fix : seg fault from sankoff cside Cause   : [caml_alloc_custom] might trigger garbage collection, and gc might move memory around, the stuff we loaded at the beginning of the funciton by Data_custom_val, might not be the same when gc is triggered. Fix     : when we call caml_alloc_custom to register a new block of memory, make sure either 1.we don't need to access old stuff loaded by Data_custom_val any more. or 2.if we do, reload them again. they are still in caml register, just could be in another place.
'1' can be acheived by caml_alloc_custom with only the pointer to the data
structure. if we do this, most human being will allocate a local block of
memory, fill it in by calling function(like median function) with old input
stuff,then at the end of function, call [caml_alloc_custom], register the
pointer poiting to the local block, return that to caml side. at this
point, the old stuff might be moved to another place, but who cares.

'2' is the painful way if we happened to do caml_alloc_custom with the whole
data structure, and maybe stuff its pointers point to. if we do this,
caml_alloc_custom will be called before median function, but old stuff might
be gone because of gc, so we have to load them again with Data_custom_val before the
median function.