This is a pre-emptive bugfix based on the numerical.ml issue in 2655
This adds a convergence criteria for the following case in brents method

A ------XM------ B

and, A<=X<=B,

Convergence happens when the distance between X and XM is greater than the
distance between A and XM + some tolerance. The problem occurs when the function
is flat in that section and no points are updated. Thus an infinite loop would
occur unless a max-iteration call was made. Because these are heavy-routines
being optimized, we need a better convergence condition. We ensure that the new
point is not the same as the previous.