In continuing modifing the likelihood model; fixed classification function
The classification function estimates a rate matrix from a set of transitions on
a tree. These are categorized for calculating relative rates. Further
optimization of the model is necessary, but it's a good start.

This change continues the previous likelihood changeset in removing one
parameter from the gtr array. this allows faster convergence, and probably more
stability in the function.