Usability - Automatically exit at script end when running in parallel; Comments
Added commends to the supports module, and improved the usability of parallel
applications when certain options are excluded from the runtime. The comments
document this change, but in short, the application needs to exit on script
completion, and if errors in the script happen. Since we know when parallel
applications are running we should enable these properties by default.