Reversion to setting cost_fn back to accepting clade
although i think in the long run this needs to be changed, there is an issue
where the clade is not a leaf as assumed by the function. I need to sit down and
determine exactly what that node means, for now we loosen the model compare, and
always select the left child, and pass this as the right child.