Bug Fix : costmatrix with non-0 diag, cost report by all_root is different from tree cost Cause : distance function return different cost than alignment function.
1. median with/without gap

say cost between same states is x, different states is y.  we have two node A
and B
A: a a b
B: a b
alignment function give us a median M
M: a/- a b
the cost come out of alignment function is 5x+y

later, to_single function will resolve the ambiguity in median and get rid of
pure gap.for example above, M become [a b]
when reporting tree cost, distance function iterate over each branch, calculate
cost between A and M, B and M
now we have
A: a a b
M: a b
B: a b
cost_A_M should be 2x+y, cost_B_M should be 2x.
they sum up to 4x+y , which is not the alignment cost we got before

2. missing data

when we have 1 out of 2 characters missing for each node
node A : [empty][AAA]
node B : [TTT][empty]
we used to return non-empty one as median, cost=0.

median will be [TTT][AAA] with cost  0 + 0 = 0
later we calculate distance between median and A, median and B,
median : [TTT][AAA]
node A : [emtpy][AAA]
this won't be 0.

this give us a alignment cost lower than distance cost.

Fix for 1: this is not bug, just the way it is. during the swap, higher
alignment cost might loose good trees. we need to deal with that later

Fix for 2: change alignment function. when one of the input is empty, return the
other sequence as median, also get cost[other sequence][other sequence] by
calling recost of sequence.Align module with original cost matrix.