Removed IFDEF statements; passing additional options to NCM code for alphabet
IFDEF statements in these modules are not necessary when they are eventually
called in the MLMODEL / MLSTATICCS / MLDYNAMIC modules.

The alphabet option is also accepted in the NCM command to specify the size of
the alphabet -- fixed, or set to the observed states, or set to the alphabet
associated to the charater; just like morphology.