Feature - Expanding *IC to allow multiple character sets.
Now we break apart the characters to allow multiple models optimized for each
character. We accomplish this by setting the weights to other characters to 0,
with our previous changeset of filtering 0 cost characters. This results in an
easy way to optimize individual characters without too much overhead.

One issue is clearing ALL weights first, and adding them back at the end of the
computation to get better output and faster computational times.