Updated Supports to deal with floating point costs and assignments
This is necessary when we apply likelihood supports based on costs.  This also
is required for more different weights in parsimony scores and more accurate
results in the bremer and consensus calculations. We report 4 decimals of
output. The change also allows us to report 'inf' instead of 'max_int' under no
support. This is more inline with actuality in the calculation, and the simple
removal if necessary.