Added interface to pnorm, qnorm, and dnorm for AU and other statistical tests.
qnorm was already included in the gamma functions as a special case
approximation for the percentage points.

dnorm is a straightforward implementation of the density function in ocaml

pnorm is taken from the statistical package R; basic comments are removed and
for further information on the function, it is advised to see the R package and
implementation for the function --only a subset of the features are exposed by
the ocaml implementation.