Added checks for setting the root
Previously we can set the root to any integer; this carries a lot of problems in
further commands. We can check that the value is a node, and do so. We also
check that the node is an actual terminal and not an internal node that would
not be something a user should really reference.