Bugfix - Fixed Hashtbl lookup for TCM matrices of multiple partitions
Symptom - lots of 'warning' messages regarding large memory usage
Cause   - Although our choice to use a hashtbl to store situations with the same
          key (the alphabet) is fine, we were never adding the new data to the
          Hashtbl when it was created.
Solution- Use the created hashtbl and insert values appropriately.

Also, we removed the 'warning' message. Although the message is helpful, it got
to a point of being obnoxious. The usage should be left to the reader to
determine if they are going to have enough memory.