Added AssignTreeNames and UniqueNames to scripting
Both of these functions are used to identify duplicated trees from the parallel
swap command to be completed. The script analyzer was modified to allow these
commands to work with the parallel swap commands parallel pipeline. This has
been tested on a number of scripts with great success.

AssignTreeNames -- assigns unique names to the trees in memory, which are
  manipulated during the swap so they can be identified later.

UniqueNames -- filters out duplications of lesser quality. This will allow the
  parallel swap command to choose the best from the candidate search space.